Organized Java Files into UI and Model Packages
