+++ app/src/main/AndroidManifest.xml
@@ -14,16 +14,16 @@
         android:theme="@style/Theme.Dogtest"
         tools:targetApi="31">
         <activity
-            android:name=".ChangePasswordActivity"
+            android:name=".ui.activity.ChangePasswordActivity"
             android:exported="false"
             android:label="@string/title_activity_password_change"
             android:theme="@style/Theme.Dogtest.NoActionBar" />
         <activity
-            android:name=".Register"
+            android:name=".ui.activity.Register"
             android:exported="false"
             android:theme="@style/Theme.Material3.Light.NoActionBar" />
         <activity
-            android:name=".Login"
+            android:name=".ui.activity.Login"
             android:exported="true"
             android:theme="@style/Theme.Material3.Light.NoActionBar">
             <intent-filter>
@@ -33,10 +33,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.activity.MainActivity"
             android:exported="true"
             android:theme="@style/NoActionBar"></activity>
-        <activity android:name=".DeleteAccountActivity"/>
+        <activity android:name=".ui.activity.DeleteAccountActivity"/>
     </application>
 
 </manifest>
