+++ app/src/main/java/com/example/dog_test/model/User.java
@@ -1,4 +1,4 @@
-package com.example.dog_test;
+package com.example.dog_test.model;
 
 public class User {
