+++ app/src/main/java/com/example/dog_test/ui/activity/ChangePasswordActivity.java
@@ -1,4 +1,4 @@
-package com.example.dog_test;
+package com.example.dog_test.ui.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -10,6 +10,7 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.dog_test.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.android.material.textfield.TextInputEditText;
