+++ app/src/main/java/com/example/dog_test/ui/activity/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.dog_test;
+package com.example.dog_test.ui.activity;
 
 import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
@@ -10,11 +10,12 @@ import android.view.MenuItem;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
-import android.view.Window;
 import android.widget.Button;
 import android.widget.TextView;
 
+import com.example.dog_test.R;
+import com.example.dog_test.ui.fragment.SettingsFragment;
+import com.example.dog_test.ui.fragment.HomeFragment;
 import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
