+++ app/src/main/java/com/example/dog_test/ui/fragment/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.example.dog_test;
+package com.example.dog_test.ui.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -8,6 +8,8 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
+import com.example.dog_test.R;
 
 public class HomeFragment extends Fragment {
