+++ app/src/main/java/com/example/dog_test/ui/fragment/SettingsFragment.java
@@ -1,7 +1,5 @@
-package com.example.dog_test;
+package com.example.dog_test.ui.fragment;
 
-import android.app.AlertDialog;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
@@ -9,12 +7,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.LinearLayout;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
+import com.example.dog_test.R;
+import com.example.dog_test.ui.activity.ChangePasswordActivity;
+import com.example.dog_test.ui.activity.DeleteAccountActivity;
 
 public class SettingsFragment extends Fragment implements View.OnClickListener {
