+++ app/src/main/res/layout/activity_delete_account.xml
@@ -7,7 +7,7 @@
     android:gravity="center"
     android:layout_height="match_parent"
     android:padding="15dp"
-    tools:context=".Register">
+    tools:context=".ui.activity.Register">
 
     <TextView
         android:layout_width="match_parent"
