+++ app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
     android:gravity="center"
     android:layout_height="match_parent"
     android:padding="15dp"
-    tools:context=".Login">
+    tools:context=".ui.activity.Login">
 
     <TextView
         android:layout_width="match_parent"
